Removed unnecessary interfaces
