+++ src/main/java/uk/gov/ons/ctp/response/casesvc/client/InternetAccessCodeSvcClient.java
@@ -1,4 +1,4 @@
-package uk.gov.ons.ctp.response.casesvc.service.impl;
+package uk.gov.ons.ctp.response.casesvc.client;
 
 import static uk.gov.ons.ctp.response.casesvc.utility.Constants.SYSTEM;
 
@@ -19,16 +19,14 @@ import org.springframework.web.client.RestTemplate;
 import org.springframework.web.util.UriComponents;
 import uk.gov.ons.ctp.common.rest.RestUtility;
 import uk.gov.ons.ctp.response.casesvc.config.AppConfig;
-import uk.gov.ons.ctp.response.casesvc.service.InternetAccessCodeSvcClientService;
 import uk.gov.ons.ctp.response.iac.representation.CreateInternetAccessCodeDTO;
 import uk.gov.ons.ctp.response.iac.representation.InternetAccessCodeDTO;
 import uk.gov.ons.ctp.response.iac.representation.UpdateInternetAccessCodeDTO;
 
 /** The impl of the service which calls the IAC service via REST */
 @Service
-public class InternetAccessCodeSvcClientServiceImpl implements InternetAccessCodeSvcClientService {
-  private static final Logger log =
-      LoggerFactory.getLogger(InternetAccessCodeSvcClientServiceImpl.class);
+public class InternetAccessCodeSvcClient {
+  private static final Logger log = LoggerFactory.getLogger(InternetAccessCodeSvcClient.class);
 
   private AppConfig appConfig;
   private RestTemplate restTemplate;
@@ -36,7 +34,7 @@ public class InternetAccessCodeSvcClientServiceImpl implements InternetAccessCod
 
   /** Constructor for InternetAccessCodeSvcClientServiceImpl */
   @Autowired
-  public InternetAccessCodeSvcClientServiceImpl(
+  public InternetAccessCodeSvcClient(
       final AppConfig appConfig,
       final RestTemplate restTemplate,
       final @Qualifier("iacServiceRestUtility") RestUtility restUtility) {
@@ -45,11 +43,16 @@ public class InternetAccessCodeSvcClientServiceImpl implements InternetAccessCod
     this.restUtility = restUtility;
   }
 
+  /**
+   * To generate IACs
+   *
+   * @param count the number of IACs to generate
+   * @return a list of IACs
+   */
   @Retryable(
       value = {RestClientException.class},
       maxAttemptsExpression = "#{${retries.maxAttempts}}",
       backoff = @Backoff(delayExpression = "#{${retries.backoff}}"))
-  @Override
   public List<String> generateIACs(int count) {
     log.debug("Generating iac codes, count: {}", count);
     UriComponents uriComponents =
@@ -66,11 +69,15 @@ public class InternetAccessCodeSvcClientServiceImpl implements InternetAccessCod
     return Arrays.asList(responseEntity.getBody());
   }
 
+  /**
+   * To disable an IAC
+   *
+   * @param iac the one to disable
+   */
   @Retryable(
       value = {RestClientException.class},
       maxAttemptsExpression = "#{${retries.maxAttempts}}",
       backoff = @Backoff(delayExpression = "#{${retries.backoff}}"))
-  @Override
   public void disableIAC(String iac) {
     log.debug("Disabling iac code");
     UriComponents uriComponents =
@@ -84,11 +91,16 @@ public class InternetAccessCodeSvcClientServiceImpl implements InternetAccessCod
     log.debug("Successfully disabled iac code");
   }
 
+  /**
+   * To determine if a given iac is active
+   *
+   * @param iac code to check active status of
+   * @return boolean true if iac is active, false if not
+   */
   @Retryable(
       value = {RestClientException.class},
       maxAttemptsExpression = "#{${retries.maxAttempts}}",
       backoff = @Backoff(delayExpression = "#{${retries.backoff}}"))
-  @Override
   public Boolean isIacActive(final String iac) {
     log.debug("Checking if iac code is active");
     UriComponents uriComponents =
