+++ src/main/java/uk/gov/ons/ctp/response/casesvc/message/CaseCreationReceiver.java
@@ -1,13 +1,28 @@
 package uk.gov.ons.ctp.response.casesvc.message;
 
+import com.godaddy.logging.Logger;
+import com.godaddy.logging.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.integration.annotation.MessageEndpoint;
+import org.springframework.integration.annotation.ServiceActivator;
 import uk.gov.ons.ctp.response.casesvc.message.sampleunitnotification.SampleUnitParent;
+import uk.gov.ons.ctp.response.casesvc.service.CaseService;
 
-/** The reader of SampleUnitParents from queue */
-public interface CaseCreationReceiver {
+/** Receive a new case from the Collection Exercise service. */
+@MessageEndpoint
+public class CaseCreationReceiver {
+  private static final Logger log = LoggerFactory.getLogger(CaseCreationReceiver.class);
+
+  @Autowired private CaseService caseService;
+
   /**
    * To process SampleUnitParents read from queue
    *
    * @param caseCreation the java representation of the message body
    */
-  void acceptSampleUnit(SampleUnitParent caseCreation);
+  @ServiceActivator(inputChannel = "caseTransformed", adviceChain = "caseRetryAdvice")
+  public void acceptSampleUnit(SampleUnitParent caseCreation) {
+    log.debug("received CaseCreation Message from queue");
+    caseService.createInitialCase(caseCreation);
+  }
 }
