+++ src/main/java/uk/gov/ons/ctp/response/casesvc/message/CaseNotificationPublisher.java
@@ -1,14 +1,31 @@
 package uk.gov.ons.ctp.response.casesvc.message;
 
+import com.godaddy.logging.Logger;
+import com.godaddy.logging.LoggerFactory;
+import org.springframework.amqp.rabbit.core.RabbitTemplate;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.integration.annotation.MessageEndpoint;
 import uk.gov.ons.ctp.response.casesvc.message.notification.CaseNotification;
 
-/** Service responsible for publishing case lifecycle events to notification channel */
-public interface CaseNotificationPublisher {
+/**
+ * Service implementation responsible for publishing case lifecycle events to notification channel
+ */
+@MessageEndpoint
+public class CaseNotificationPublisher {
+  private static final Logger log = LoggerFactory.getLogger(CaseNotificationPublisher.class);
+
+  @Qualifier("caseNotificationRabbitTemplate")
+  @Autowired
+  private RabbitTemplate rabbitTemplate;
 
   /**
    * To put one CaseNotification on the outbound channel caseNotificationOutbound
    *
    * @param caseNotification the CaseNotification to put on the outbound channel
    */
-  void sendNotification(CaseNotification caseNotification);
+  public void sendNotification(CaseNotification caseNotification) {
+    log.debug("Entering sendNotification with CaseNotification {}", caseNotification);
+    rabbitTemplate.convertAndSend(caseNotification);
+  }
 }
