+++ src/main/java/uk/gov/ons/ctp/response/casesvc/message/EventPublisher.java
@@ -1,6 +1,22 @@
 package uk.gov.ons.ctp.response.casesvc.message;
 
-public interface EventPublisher {
+import com.godaddy.logging.Logger;
+import com.godaddy.logging.LoggerFactory;
+import org.springframework.amqp.rabbit.core.RabbitTemplate;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.integration.annotation.MessageEndpoint;
 
-  public void publishEvent(String event);
+@MessageEndpoint
+public class EventPublisher {
+  private static final Logger log = LoggerFactory.getLogger(EventPublisher.class);
+
+  @Qualifier("amqpTemplate")
+  @Autowired
+  private RabbitTemplate rabbitTemplate;
+
+  public void publishEvent(String event) {
+    log.debug("Publish Event", event);
+    rabbitTemplate.convertAndSend(event);
+  }
 }
