--- src/main/java/uk/gov/ons/ctp/response/casesvc/message/impl/EventPublisherImpl.java
@@ -1,24 +1,0 @@
-package uk.gov.ons.ctp.response.casesvc.message.impl;
-
-import com.godaddy.logging.Logger;
-import com.godaddy.logging.LoggerFactory;
-import org.springframework.amqp.rabbit.core.RabbitTemplate;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.integration.annotation.MessageEndpoint;
-import uk.gov.ons.ctp.response.casesvc.message.EventPublisher;
-
-@MessageEndpoint
-public class EventPublisherImpl implements EventPublisher {
-  private static final Logger log = LoggerFactory.getLogger(EventPublisherImpl.class);
-
-  @Qualifier("amqpTemplate")
-  @Autowired
-  private RabbitTemplate rabbitTemplate;
-
-  @Override
-  public void publishEvent(String event) {
-    log.debug("Publish Event", event);
-    rabbitTemplate.convertAndSend(event);
-  }
-}
