+++ src/main/java/uk/gov/ons/ctp/response/casesvc/service/CaseReportService.java
@@ -1,7 +1,29 @@
 package uk.gov.ons.ctp.response.casesvc.service;
 
-/** The service to create reports */
-public interface CaseReportService {
-  /** To create reports */
-  void createReport();
+import com.godaddy.logging.Logger;
+import com.godaddy.logging.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+import uk.gov.ons.ctp.response.casesvc.domain.repository.CaseReportRepository;
+
+/**
+ * A CaseReportService implementation which encapsulates all business logic operating on the
+ * CaseReport entity model.
+ */
+@Service
+public class CaseReportService {
+  private static final Logger log = LoggerFactory.getLogger(CaseReportService.class);
+
+  @Autowired private CaseReportRepository caseReportRepository;
+
+  /** Creates a Report */
+  public void createReport() {
+    log.debug("Entering createReport...");
+
+    boolean reportResult = caseReportRepository.chasingReportStoredProcedure();
+    log.debug("Just ran the chasing report and result is {}", reportResult);
+
+    reportResult = caseReportRepository.caseEventsReportStoredProcedure();
+    log.debug("Just ran the case events report and result is {}", reportResult);
+  }
 }
