+++ src/test/java/uk/gov/ons/ctp/response/casesvc/client/InternetAccessCodeSvcClientImplTest.java
@@ -1,4 +1,4 @@
-package uk.gov.ons.ctp.response.casesvc.service.impl;
+package uk.gov.ons.ctp.response.casesvc.client;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
@@ -36,7 +36,7 @@ import uk.gov.ons.ctp.response.iac.representation.UpdateInternetAccessCodeDTO;
 
 /** Testing InternetAccessCodeSvcClientServiceImpl */
 @RunWith(MockitoJUnitRunner.class)
-public class InternetAccessCodeSvcClientServiceImplTest {
+public class InternetAccessCodeSvcClientImplTest {
 
   private static final String HTTP = "http";
   private static final String LOCALHOST = "localhost";
@@ -49,7 +49,7 @@ public class InternetAccessCodeSvcClientServiceImplTest {
   @Mock private RestTemplate restTemplate;
   @Mock private RestUtility restUtility;
 
-  @InjectMocks private InternetAccessCodeSvcClientServiceImpl iacSvcClientService;
+  @InjectMocks private InternetAccessCodeSvcClient internetAccessCodeSvcClient;
 
   /** Set up Mockito for tests */
   @Before
@@ -87,7 +87,7 @@ public class InternetAccessCodeSvcClientServiceImplTest {
             any(URI.class), any(HttpMethod.class), any(HttpEntity.class), any(Class.class)))
         .thenReturn(responseEntity);
 
-    List<String> codes = iacSvcClientService.generateIACs(count);
+    List<String> codes = internetAccessCodeSvcClient.generateIACs(count);
     assertTrue(codes != null);
     assertEquals(3, codes.size());
     assertTrue(codes.containsAll(Arrays.asList(new String[] {"1", "2", "3"})));
@@ -122,7 +122,7 @@ public class InternetAccessCodeSvcClientServiceImplTest {
     when(restUtility.createHttpEntity(any(UpdateInternetAccessCodeDTO.class)))
         .thenReturn(httpEntity);
 
-    iacSvcClientService.disableIAC(IAC);
+    internetAccessCodeSvcClient.disableIAC(IAC);
 
     verify(restUtility, times(1)).createUriComponents(IAC_PUT_PATH, null, IAC);
     verify(restUtility, times(1)).createHttpEntity(eq(updateInternetAccessCodeDTO));
@@ -157,7 +157,7 @@ public class InternetAccessCodeSvcClientServiceImplTest {
             uriComponents.toUri(), HttpMethod.GET, null, InternetAccessCodeDTO.class))
         .thenReturn(responseEntity);
 
-    Boolean isActive = iacSvcClientService.isIacActive(IAC);
+    Boolean isActive = internetAccessCodeSvcClient.isIacActive(IAC);
 
     assertEquals(isActive, true);
     verify(restTemplate, times(1))
