+++ src/test/java/uk/gov/ons/ctp/response/casesvc/service/impl/CaseCreationServiceTest.java
@@ -24,13 +24,14 @@ import uk.gov.ons.ctp.response.casesvc.message.sampleunitnotification.SampleUnit
 import uk.gov.ons.ctp.response.casesvc.message.sampleunitnotification.SampleUnitParent;
 import uk.gov.ons.ctp.response.casesvc.representation.CaseGroupStatus;
 import uk.gov.ons.ctp.response.casesvc.representation.CaseState;
+import uk.gov.ons.ctp.response.casesvc.service.CaseService;
 import uk.gov.ons.ctp.response.sample.representation.SampleUnitDTO;
 
 /** Test Case created by Sample */
 @RunWith(MockitoJUnitRunner.class)
 public class CaseCreationServiceTest {
 
-  @InjectMocks private CaseServiceImpl caseService;
+  @InjectMocks private CaseService caseService;
 
   @Mock private CaseRepository caseRepo;
   @Mock private CaseGroupRepository caseGroupRepo;
